+++ cmp/grid/impl/InitPersist.ts
@@ -6,6 +6,7 @@
  */
 import {PersistableState, PersistenceProvider} from '@xh/hoist/core';
 import {isEqual, isObject} from 'lodash';
+import {runInAction} from 'mobx';
 import {GridModel} from '../GridModel';
 import {ColumnState, GridModelPersistOptions} from '../Types';
 
@@ -36,7 +37,18 @@ export function initPersist(
             target: {
                 getPersistableState: () =>
                     new PersistableColumnState(gridModel.persistableColumnState),
-                setPersistableState: ({value}) => gridModel.setColumnState(value)
+                setPersistableState: ({value}) =>
+                    runInAction(() => {
+                        // Set columnState directly since GridModel.setColumnState will merge the
+                        // provided state with the current state, which is not what we want here.
+                        gridModel.columnState = gridModel.cleanColumnState(value);
+                        if (gridModel.autosizeOptions.mode === 'managed') {
+                            const columns = gridModel.columnState
+                                .filter(it => !it.manuallySized)
+                                .map(it => it.colId);
+                            gridModel.autosizeAsync({columns});
+                        }
+                    })
             },
             owner: gridModel
         });